Move logo styling and components to own folder
