+++ components/logo/LogoMidden.tsx
@@ -1,4 +1,5 @@
 import {FC, PropsWithChildren} from 'react';
+import './logo.css'
 
 const LogoMidden: FC<PropsWithChildren> = ({children}) => {
     return (
